Use hashicorp/go-retryablehttp client
